Fix some 64-bit issues in the JIT.
The macOS ARM64 JIT engine will need to have the mapping hosted
>4GB, so this fixes some 64-bit issues exposed.

Also, avoid a place that requested RWX permissions for the ARM64
JIT mapping, which is not needed, but is also a no-no on macOS
ARM64.

+++ asm/asm_opcodes.h
@@ -5,8 +5,8 @@
 
 struct asm_uop {
   int32_t uopcode;
-  int32_t value1;
-  int32_t value2;
+  intptr_t value1;
+  intptr_t value2;
   int is_eliminated;
   int is_merged;
   int32_t backend_tag;
